fixing add image

+++ BE/src/controllers/Kebijakan.js
@@ -223,56 +223,6 @@ export const deleteKebijakan = async (req, res) => {
     }
 }
 
-export const uploadImageKebijakanBaru = async (req, res) => {
-    const kebijakan = await Kebijakans.findOne({
-        where: {
-            kid: req.params.id
-        }
-    });
-
-    if (!kebijakan) {
-        return res.status(404).json({msg: "Tidak ditemukan kebijakan, harap buat kebijakan terlebih dahulu"});
-    }
-
-    let fileName = "";
-    if(req.files === null || req.files === undefined) {
-        fileName = kebijakan.foto_data;
-    } else {
-        const file = req.files.foto;
-        const fileSize = file.data.length;
-        const ext = path.extname(file.name);
-        fileName = file.md5 + ext;
-        const allowedType = ['.png','.jpg','.jpeg'];
-
-        if(!allowedType.includes(ext.toLowerCase())) {
-            return res.status(422).json({msg: "Foto yang anda masukkan tidak valid"});
-        }
-
-        if(fileSize > 50000000) {
-            return res.status(422).json({msg: "Ukuran foto harus kurang dari 50 MB"});
-        }
-
-        file.mv(`./public/images/kebijakan/${fileName}`, (err)=>{
-            if(err) {
-                return res.status(500).json({msg: err.message})
-            };
-        });
-    }
-
-    const url = `${req.protocol}://${req.get("host")}/images/kebijakan/${fileName}`;
-    
-    try {
-        await Kebijakans.update(
-            { foto_data: fileName, foto_url: url },
-            { where: { kid: req.params.id } }
-        );
-        res.status(200).json({msg: "Sukses memasukkan foto kebijakan"});
-    } catch (error) {
-        console.log(error.message);
-        res.status(400).json({msg: error.message});
-    }
-}
-
 export const editUploadImageKebijakan = async (req, res) => {
     const kebijakan = await Kebijakans.findOne({
         where: {
@@ -302,9 +252,11 @@ export const editUploadImageKebijakan = async (req, res) => {
             return res.status(422).json({msg: "Ukuran foto harus kurang dari 50 MB"});
         }
 
-        // const filepath = `./public/images/${product.image}`;
-        const filepath = `./public/images/kebijakan/${kebijakan.foto_data}`;
-        fs.unlinkSync(filepath);
+        const fotodata  = kebijakan.foto_data === "" && kebijakan.foto_url === "";
+        if (!fotodata){
+            const filepath = `./public/images/kebijakan/${kebijakan.foto_data}`;
+            fs.unlinkSync(filepath);
+        }
 
         file.mv(`./public/images/kebijakan/${fileName}`, (err)=>{
             if(err) {
@@ -312,7 +264,7 @@ export const editUploadImageKebijakan = async (req, res) => {
             };
         });
     }
-    // const name = req.body.title;
+
     const url = `${req.protocol}://${req.get("host")}/images/kebijakan/${fileName}`;
     
     try {
